# openapi/historical-pool-quote-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Historical Pool Quote API",
    "description": "Authenticated historical directed pool quotes from reconstructed Base state.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "1"
  },
  "paths": {
...
    "/jobs/{jobId}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "operationId": "get_job",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "description": "Opaque UUID version 4 job ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-4[0-9a-fA-F]{3}-[89aAbB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
            }
          },
          {
            "name": "X-DSolver-History-API-Revision",
            "in": "header",
            "description": "Exact API revision",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 1,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Current job envelope",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "format": "int64",
                  "minimum": 0
                },
                "description": "Present while work is incomplete"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Job ID or API revision is invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token is absent or wrong",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Job is not retained",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/jobs/{jobId}/cancel": {
      "post": {
        "tags": [
          "jobs"
        ],
        "operationId": "cancel_job",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "description": "Opaque UUID version 4 job ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-4[0-9a-fA-F]{3}-[89aAbB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
            }
          },
          {
            "name": "X-DSolver-History-API-Revision",
            "in": "header",
            "description": "Exact API revision",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 1,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Current state after the cancellation request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Job ID or API revision is invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token is absent or wrong",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Job is not retained",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "description": "Safe error returned by the HTTP API.",
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ApiErrorCode"
          },
          "details": {},
          "message": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "ApiErrorCode": {
        "type": "string",
        "description": "Stable code returned by an HTTP admission or polling error.",
        "enum": [
          "invalid_request",
          "unauthorized",
          "job_not_found",
          "request_id_conflict",
          "comparison_budget_exceeded",
          "queue_full",
          "service_unavailable"
        ]
      },
      "Backend": {
        "type": "string",
        "description": "Simulator state partition used by a historical request.",
        "enum": [
          "native",
          "vm",
          "rfq"
        ]
      },
...
      "ConsistencyCheckProgress": {
        "type": "object",
        "description": "Consistency progress counts completed checkpoint pairs.",
        "required": [
          "completedCheckpointPairs",
          "totalCheckpointPairs",
          "percentComplete"
        ],
        "properties": {
          "completedCheckpointPairs": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "percentComplete": {
            "type": "integer",
            "format": "int32",
            "maximum": 100,
            "minimum": 0
          },
          "totalCheckpointPairs": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        },
        "additionalProperties": false
      },
      "ConsistencyCheckReport": {
        "type": "object",
        "description": "Completed report for a historical state consistency check.",
        "required": [
          "requestId",
          "chainId",
          "backends",
          "selection",
          "selectionStatus",
          "storage",
          "summary",
          "pairs"
        ],
        "properties": {
          "backends": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Backend"
            }
          },
          "chainId": {
            "type": "integer",
            "format": "int64",
            "maximum": 8453,
            "minimum": 8453
          },
          "pairs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConsistencyPairReport"
            }
          },
          "requestId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-4[0-9a-fA-F]{3}-[89aAbB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
          },
          "selection": {
            "$ref": "#/components/schemas/ConsistencySelection"
          },
          "selectionStatus": {
            "$ref": "#/components/schemas/ConsistencySelectionStatus"
          },
          "storage": {
            "$ref": "#/components/schemas/StorageVerificationSummary"
          },
          "summary": {
            "$ref": "#/components/schemas/ConsistencySummary"
          }
        },
        "additionalProperties": false
      },
...
      "ConsistencySelectionStatus": {
        "type": "string",
        "description": "Selection-level status for a consistency check.",
        "enum": [
          "compared",
          "notComparable"
        ]
      },
...
      "GapReference": {
        "type": "object",
        "description": "Response-local gap reference in a quote result.",
        "required": [
          "ref",
          "kind"
        ],
        "properties": {
          "cause": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/GapCause"
              }
            ]
          },
          "fromBlock": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "fromObservedAtMs": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "fromPosition": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StreamPosition"
              }
            ]
          },
          "kind": {
            "$ref": "#/components/schemas/GapKind"
          },
          "ref": {
            "type": "string"
          },
          "toBlockInclusive": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "toObservedAtMs": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "toPosition": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StreamPosition"
              }
            ]
          }
        },
        "additionalProperties": false
      },
...
      "HistoricalQuoteProgress": {
        "type": "object",
        "description": "Quote progress counts completed lag comparisons.",
        "required": [
          "completedComparisons",
          "totalComparisons",
          "percentComplete"
        ],
        "properties": {
          "completedComparisons": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "percentComplete": {
            "type": "integer",
            "format": "int32",
            "maximum": 100,
            "minimum": 0
          },
          "totalComparisons": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        },
        "additionalProperties": false
      },
      "HistoricalQuoteResult": {
        "type": "object",
        "description": "Result for a completed historical quote job.",
        "required": [
          "requestId",
          "chainId",
          "pool",
          "visibleThroughBlock",
          "gaps",
          "results"
        ],
        "properties": {
          "chainId": {
            "type": "integer",
            "format": "int64",
            "maximum": 8453,
            "minimum": 8453
          },
          "gaps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GapReference"
            }
          },
          "pool": {
            "$ref": "#/components/schemas/PoolSelector"
          },
          "requestId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-4[0-9a-fA-F]{3}-[89aAbB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HistoricalQuoteEntry"
            }
          },
          "visibleThroughBlock": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        },
        "additionalProperties": false
      },
...
      "JobEnvelope": {
        "type": "object",
        "description": "Current job state returned by polling or cancellation.",
        "required": [
          "jobId",
          "requestId",
          "jobType",
          "state",
          "submittedAt",
          "deadlineAt",
          "cancellationRequested",
          "progress"
        ],
        "properties": {
          "cancellationRequested": {
            "type": "boolean"
          },
          "deadlineAt": {
            "type": "string",
            "format": "date-time"
          },
          "failure": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/JobFailure"
              }
            ]
          },
          "finishedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "jobId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-4[0-9a-fA-F]{3}-[89aAbB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
          },
          "jobType": {
            "$ref": "#/components/schemas/JobType"
          },
          "progress": {
            "$ref": "#/components/schemas/JobProgress"
          },
          "queuePosition": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 1
          },
          "requestId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-4[0-9a-fA-F]{3}-[89aAbB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
          },
          "result": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/JobResult"
              }
            ]
          },
          "startedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "state": {
            "$ref": "#/components/schemas/JobState"
          },
          "submittedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "JobFailure": {
        "type": "object",
        "description": "Safe failure stored in a terminal job envelope.",
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/JobFailureCode"
          },
          "details": {},
          "message": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "JobFailureCode": {
        "type": "string",
        "description": "Stable terminal failure code.",
        "enum": [
          "deadline_exceeded",
          "history_read_failed",
          "historical_data_invalid",
          "state_reconstruction_failed",
          "check_budget_exceeded",
          "history_not_yet_stored",
          "history_unavailable",
          "read_limit_exceeded",
          "internal_error"
        ]
      },
      "JobProgress": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/HistoricalQuoteProgress"
          },
          {
            "$ref": "#/components/schemas/ConsistencyCheckProgress"
          },
          {
            "$ref": "#/components/schemas/RawHistoryProgress"
          }
        ],
        "description": "Progress shape selected by the job type."
      },
      "JobResult": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/HistoricalQuoteResult"
          },
          {
            "$ref": "#/components/schemas/ConsistencyCheckReport"
          },
          {
            "$ref": "#/components/schemas/RawSnapshotResult"
          },
          {
            "$ref": "#/components/schemas/StoredMessagesResult"
          }
        ],
        "description": "Terminal result selected by the job type."
      },
      "JobState": {
        "type": "string",
        "description": "Lifecycle state of an admitted job.",
        "enum": [
          "queued",
          "running",
          "completed",
          "failed",
          "cancelled"
        ]
      },
...
      "JobType": {
        "type": "string",
        "description": "Work type stored by the shared in-memory scheduler.",
        "enum": [
          "historicalQuote",
          "historicalStateConsistencyCheck",
          "rawSnapshot",
          "storedMessages"
        ]
      },
...
      "PoolSelector": {
        "type": "object",
        "description": "Exact pool and token direction selected by the caller.",
        "required": [
          "backend",
          "protocol",
          "componentId",
          "tokenIn",
          "tokenOut"
        ],
        "properties": {
          "backend": {
            "$ref": "#/components/schemas/Backend"
          },
          "componentId": {
            "type": "string",
            "minLength": 1
          },
          "protocol": {
            "type": "string",
            "minLength": 1
          },
          "tokenIn": {
            "type": "string",
            "minLength": 1
          },
          "tokenOut": {
            "type": "string",
            "minLength": 1
          }
        },
        "additionalProperties": false
      },
...
      "RawHistoryProgress": {
        "type": "object",
        "description": "Raw-history progress, a single read that is either pending or done.",
        "required": [
          "percentComplete"
        ],
        "properties": {
          "percentComplete": {
            "type": "integer",
            "format": "int32",
            "maximum": 100,
            "minimum": 0
          }
        },
        "additionalProperties": false
      },
...
      "RawSnapshotResult": {
        "type": "object",
        "description": "The raw snapshot at a stream position.",
        "required": [
          "position",
          "checkpointPosition",
          "partitions"
        ],
        "properties": {
          "checkpointPosition": {
            "$ref": "#/components/schemas/StreamPosition",
            "description": "The checkpoint the snapshot was rebuilt from."
          },
          "partitions": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "description": "One snapshot partition per requested backend, in the broadcaster's\nsnapshot wire form."
          },
          "position": {
            "$ref": "#/components/schemas/StreamPosition"
          }
        },
        "additionalProperties": false
      },
...
      "StoredMessagesResult": {
        "type": "object",
        "description": "The stored messages in a range, in stream order.",
        "required": [
          "after",
          "through",
          "messages"
        ],
        "properties": {
          "after": {
            "$ref": "#/components/schemas/StreamPosition"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StoredMessage"
            }
          },
          "through": {
            "$ref": "#/components/schemas/StreamPosition"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}